chore: add support for caching `node_modules` within a nested repository (AztecProtocol/aztec-packages#12862)
This PR removes a hack added in AztecProtocol/aztec-packages#12760

This fixes an issue where we were querying the root git repository for
hashes for files which are only tracked in the `noir/noir-repo`
repository. We now set the `REPO_PATH` env variable so that we run `git
ls-tree` on the correct repository.