chore: node 22 (AztecProtocol/aztec-packages#14248)
Updated Node to the latest LTS (22.15.0 at the time of writing). Most of
of the diff is due to the forced update to ESlint 9 and Prettier 3,
which had many format changes. However, there's a few good things coming
with these:

- Prettier has more consistent constructor formatting
- ESLint now detects many more instances of hanging promises and
improper stringification of objects
- We got rid of the now unsupported type assertions (`assert { type:
'json' }`) in favor of import attributes (`with { type: 'json' }`)

---------

Co-authored-by: Charlie Lye <5764343+charlielye@users.noreply.github.com>